+++ NEWS.rst
@@ -20,7 +20,7 @@ NEWS for system-image updater
    (LP: #1238290)
  * Remove all references to the `[system]threads` variable since it is no
    longer used, after the integration of the download manager.
- * Through the use of psutils library, re-enable some previously skipped
+ * Through the use of the psutil library, re-enable some previously skipped
    tests.  (LP: #1206588)
 
 1.8 (2013-10-02)